requirements.txt
Flask==3.0.3
numpy==1.26.4
Pillow==10.4.0
tensorflow==2.15.0
scipy==1.11.4
Gunicorn